LLM/LLMsTrainer/requirements.txt
absl_py==1.4.0
datasets==2.10.1
deepspeed==0.9.2
matplotlib==3.7.1
numpy==1.24.3
peft==0.3.0
psutil==5.9.5
PyYAML==6.0
rich==13.3.5
sentencepiece==0.1.99
torch==2.0.0
torchinfo==1.7.2
tqdm==4.65.0
transformers==4.29.1
zstandard==0.21.0
xformers==0.0.19
streamlit==1.24.1
streamlit-chat==0.0.2.2
plotly==5.11.0